Refactor Initialisation
